+++ crates/fuzz/fuzz_targets/accounts_call.rs
@@ -10,7 +10,8 @@ use rand::{RngCore, SeedableRng};
 use sov_accounts::{AccountConfig, AccountData, Accounts, CallMessage};
 use sov_modules_api::capabilities::mocks::MockKernel;
 use sov_modules_api::{
-    Context, CredentialId, DaSpec, Module, PrivateKey, PublicKey, Spec, StateCheckpoint, WorkingSet,
+    Context, CredentialId, DaSpec, ExecutionContext, Module, PrivateKey, PublicKey, SequencerType,
+    Spec, StateCheckpoint, WorkingSet,
 };
 use sov_test_utils::storage::SimpleStorageManager;
 use sov_test_utils::TestPrivateKey;
@@ -87,8 +88,15 @@ fuzz_target!(
         for _ in 0..iterations {
             // we use slices for better select performance
             let sender = addresses.choose(rng).unwrap();
-            let context =
-                Context::<S>::new(*sender, Default::default(), sequencer, sequencer_da, None);
+            let context = Context::<S>::new(
+                *sender,
+                Default::default(),
+                sequencer,
+                sequencer_da,
+                None,
+                ExecutionContext::Node,
+                SequencerType::Preferred,
+            );
 
             // clear previous state
             let previous = state.get(sender).unwrap().as_hex();