+++ crates/module-system/sov-capabilities/src/lib.rs
@@ -14,6 +14,7 @@ use sov_modules_api::transaction::{
     AuthenticatedTransactionData, ProverReward, RemainingFunds, SequencerReward,
 };
 use sov_modules_api::ExecutionContext;
+use sov_modules_api::SequencerType;
 use sov_modules_api::{
     AggregatedProofPublicData, Amount, Context, DaSpec, Gas, GetGasPrice, InfallibleStateAccessor,
     InvalidProofError, ModuleInfo, OperatingMode, Rewards, SovAttestation,
@@ -267,6 +268,8 @@ impl<S: Spec, T> TransactionAuthorizer<S> for StandardProvenRollupCapabilities<'
         sequencer_rollup_address: S::Address,
         state: &mut impl StateAccessor,
         sequencing_data: Option<Bytes>,
+        execution_context: ExecutionContext,
+        sequencer_type: SequencerType,
     ) -> anyhow::Result<Context<S>> {
         // This should be resolved by the sequencer registry during blob selection
         let sender = self.accounts.resolve_sender_address(
@@ -280,6 +283,8 @@ impl<S: Spec, T> TransactionAuthorizer<S> for StandardProvenRollupCapabilities<'
             sequencer_rollup_address,
             *sequencer,
             sequencing_data,
+            execution_context,
+            sequencer_type,
         ))
     }
 
@@ -288,6 +293,7 @@ impl<S: Spec, T> TransactionAuthorizer<S> for StandardProvenRollupCapabilities<'
         auth_data: &AuthorizationData<S>,
         sequencer: &<<S as Spec>::Da as DaSpec>::Address,
         state: &mut impl StateAccessor,
+        execution_context: ExecutionContext,
     ) -> anyhow::Result<Context<S>> {
         let sender = self.accounts.resolve_sender_address(
             &auth_data.default_address,
@@ -301,6 +307,8 @@ impl<S: Spec, T> TransactionAuthorizer<S> for StandardProvenRollupCapabilities<'
             sender,
             *sequencer,
             None,
+            execution_context,
+            SequencerType::NonPreferred,
         ))
     }
 }