+++ crates/module-system/sov-modules-api/src/module/spec.rs
@@ -7,6 +7,7 @@ use serde::{Deserialize, Serialize};
 use sov_rollup_interface::crypto::{CredentialId, Signature};
 use sov_rollup_interface::da::DaSpec;
 use sov_rollup_interface::optimistic::Attestation;
+use sov_rollup_interface::stf::ExecutionContext;
 use sov_rollup_interface::zk::{CryptoSpec, StateTransitionPublicData, Zkvm};
 use sov_rollup_interface::{BasicAddress, Bytes};
 use sov_state::{Storage, StorageProof};
@@ -14,7 +15,7 @@ use sov_state::{Storage, StorageProof};
 use crate::gas::Gas;
 use crate::higher_kinded_types::Generic;
 use crate::transaction::Credentials;
-use crate::{PublicKeyExt, SignatureExt};
+use crate::{PublicKeyExt, SequencerType, SignatureExt};
 
 /// The `Spec` trait configures certain key primitives to be used by a by a particular instance of a rollup.
 /// `Spec` is almost always implemented on a Context object; since all Modules are generic
@@ -164,6 +165,10 @@ pub struct Context<S: Spec> {
     sequencing_data: Option<Bytes>,
     /// The rollup address that pays the gas fees for the transaction.
     gas_refund_recipient: S::Address,
+    /// The execution context of the transaction.
+    execution_context: ExecutionContext,
+    /// The type of sequencer that published the transaction.
+    sequencer_type: SequencerType,
 }
 
 impl<S: Spec> Context<S> {
@@ -197,6 +202,21 @@ impl<S: Spec> Context<S> {
         self.gas_refund_recipient = recipient;
     }
 
+    /// Returns the execution context of the transaction.
+    pub fn execution_context(&self) -> ExecutionContext {
+        self.execution_context
+    }
+
+    /// Returns whether the sequencer of the current batch is the preferred sequencer.
+    pub fn sequencer_is_preferred(&self) -> bool {
+        self.sequencer_type == SequencerType::Preferred
+    }
+
+    /// Returns the type of sequencer that published the transaction.
+    pub fn sequencer_type(&self) -> SequencerType {
+        self.sequencer_type
+    }
+
     /// Constructs a new Context with the provided sender as the payer.
     pub fn new(
         sender: S::Address,
@@ -204,6 +224,8 @@ impl<S: Spec> Context<S> {
         sequencer: S::Address,
         sequencer_da_address: <S::Da as DaSpec>::Address,
         sequencing_data: Option<Bytes>,
+        execution_context: ExecutionContext,
+        sequencer_type: SequencerType,
     ) -> Self {
         Self::with_payer(
             sender,
@@ -212,6 +234,8 @@ impl<S: Spec> Context<S> {
             sequencer_da_address,
             sender,
             sequencing_data,
+            execution_context,
+            sequencer_type,
         )
     }
 
@@ -223,6 +247,8 @@ impl<S: Spec> Context<S> {
         sequencer_da_address: <S::Da as DaSpec>::Address,
         payer: S::Address,
         sequencing_data: Option<Bytes>,
+        execution_context: ExecutionContext,
+        sequencer_type: SequencerType,
     ) -> Self {
         Self {
             sender_credentials,
@@ -231,6 +257,8 @@ impl<S: Spec> Context<S> {
             sequencer_da_address,
             gas_refund_recipient: payer,
             sequencing_data,
+            execution_context,
+            sequencer_type,
         }
     }
 
@@ -256,8 +284,9 @@ pub type SovStateTransitionPublicData<S> = StateTransitionPublicData<
 
 #[cfg(feature = "arbitrary")]
 mod arbitrary {
+    use crate::common::SequencerType;
     use ::arbitrary::{Arbitrary, Unstructured};
-    use sov_rollup_interface::da::DaSpec;
+    use sov_rollup_interface::{da::DaSpec, stf::ExecutionContext};
 
     use super::{Context, Spec};
 
@@ -278,6 +307,8 @@ mod arbitrary {
                 sequencer,
                 sequencer_da_address,
                 None,
+                ExecutionContext::Node,
+                SequencerType::NonPreferred,
             ))
         }
     }